affixable memory cache
